Fix drain_from_back
In drain_from_back, we want to add elements to the front of the chunk. If there
is not enough space on the front/left side, then we first need to shift the data
to the right.

This commit makes sure that we shift that data to the very end of the chunk
and that the left offset is set to the chunk size minus length of the data,
in order to maintain that length = right - left.