Additional traits for reference iterators
Most importantly, Clone. See https://github.com/bodil/im-rs/issues/134.

Piggy-backing Debug when possible, because, why not.

Not implementing Copy even where it would be possible, as that could
turn into footgun (iterators "forking").

Not implementing Clone on consuming iterators, as that also clones the
elements and that's probably not intended.